requirements.txt
alembic==1.13.1
APScheduler==3.10.4
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cssmin==0.2.0
cryptography
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
eventlet==0.36.1
filetype==1.2.0
Flask==3.0.0
Flask-Admin==1.6.1
Flask-APScheduler==1.13.1
Flask-Assets==2.1.0
Flask-Babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.2.0
Flask-Cors==4.0.1
Flask-Dance==7.1.0
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Session==0.6.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-SSLify==0.1.5
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
flask-mailman==1.1.1
greenlet==3.0.3
gunicorn
html5lib==1.1
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsmin==3.0.1
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.1
Pillow==10.4.0
platformdirs==4.2.2
PyJWT==2.8.0
PyOpenSSL==24.1.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
pytz==2024.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.21
stripe==8.6.0
tomli==2.0.1
tzlocal==5.2
urllib3==2.2.2
visitor==0.1.3
waitress
webassets==2.0
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
whitenoise
psycopg2-binary